# contracts/api/openapi.yaml
openapi: "3.1.0"
info:
  title: "tok-friends API"
  version: "1.0.0"
  description: "Auto-generated contract based on NestJS controllers."
servers:
  -
    url: "https://api.tokfriends.app"
    description: "Production (example)"
  -
    url: "http://localhost:4000"
    description: "Local"
components:
  schemas:
...
    UserSummary:
      type: "object"
      properties:
        id:
          type: "string"
        email:
          type:
            - "string"
            - "null"
        displayName:
          type:
            - "string"
            - "null"
        status:
          type: "string"
        provider:
          type:
            - "string"
            - "null"
        pointsBalance:
          type: "integer"
        createdAt:
          type: "string"
        profile:
          type:
            - "object"
            - "null"
          additionalProperties: true
        counts:
          type:
            - "object"
            - "null"
          additionalProperties: true
    UsersMeResponse:
      type: "object"
      properties:
        ok:
          type: "boolean"
        data:
          $ref: "#/components/schemas/UserSummary"
    UsersUpdateRequest:
      type: "object"
      properties:
        displayName:
          type: "string"
        nickname:
          type: "string"
        bio:
          type: "string"
        region1:
          type: "string"
        region2:
          type: "string"
        interests:
          oneOf:
            -
              type: "array"
              items:
                type: "string"
            -
              type: "string"
        marketingOptIn:
          oneOf:
            -
              type: "boolean"
            -
              type: "string"
        headline:
          type: "string"
        avatarUri:
          type: "string"
    PublicUserSearchItem:
      type: "object"
      properties:
        id:
          type: "string"
        displayName:
          type:
            - "string"
            - "null"
        nickname:
          type:
            - "string"
            - "null"
        bio:
          type:
            - "string"
            - "null"
        interests:
          type: "array"
          items:
            type: "string"
    UsersSearchResponse:
      type: "object"
      properties:
        ok:
          type: "boolean"
        data:
          type: "array"
          items:
            $ref: "#/components/schemas/PublicUserSearchItem"
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/PublicUserSearchItem"
    PublicUserProfile:
      type: "object"
      properties:
        id:
          type: "string"
        displayName:
          type:
            - "string"
            - "null"
        region1:
          type:
            - "string"
            - "null"
        region2:
          type:
            - "string"
            - "null"
        profile:
          type:
            - "object"
            - "null"
          properties:
            nickname:
              type:
                - "string"
                - "null"
            bio:
              type:
                - "string"
                - "null"
            headline:
              type:
                - "string"
                - "null"
            avatarUri:
              type:
                - "string"
                - "null"
            interests:
              type: "array"
              items:
                type: "string"
            badges:
              type: "array"
              items:
                type: "string"
    PublicUserResponse:
      type: "object"
      properties:
        ok:
          type: "boolean"
        data:
          $ref: "#/components/schemas/PublicUserProfile"
...
    AdminUserDetailResponse:
      type: "object"
      properties:
        ok:
          type: "boolean"
        data:
          type: "object"
          additionalProperties: true
    AdminUserProfileUpdate:
      type: "object"
      additionalProperties: true
...
    CreatePostRequest:
      type: "object"
      required:
        - "topicId"
        - "content"
      properties:
        topicId:
          type: "string"
        content:
          type: "string"
    Post:
      type: "object"
      properties:
        id:
          type: "string"
        userId:
          type: "string"
        topicId:
          type: "string"
        content:
          type: "string"
        createdAt:
          type: "string"
    PaginatedPosts:
      type: "object"
      properties:
        items:
          type: "array"
          items:
            $ref: "#/components/schemas/Post"
        nextCursor:
          type:
            - "string"
            - "null"
        hasMore:
          type: "boolean"
...
    CommunityReport:
      type: "object"
      required:
        - "reason"
      properties:
        targetUserId:
          type: "string"
        postId:
          type: "string"
        reason:
          type: "string"
    CommunityBlock:
      type: "object"
      required:
        - "blockedUserId"
      properties:
        blockedUserId:
          type: "string"
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
paths:
...
  /users/me:
    get:
      summary: "Get current user"
      tags:
        - "users"
      responses:
        200:
          description: "Current profile"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersMeResponse"
        404:
          description: "Not found"
      security:
        -
          bearerAuth:
  /users/search:
    get:
      summary: "Search users"
      tags:
        - "users"
      responses:
        200:
          description: "Matches"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersSearchResponse"
      security:
        -
          bearerAuth:
  /users/{id}:
    patch:
      summary: "Update profile"
      tags:
        - "users"
      responses:
        200:
          description: "Updated profile"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersMeResponse"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      security:
        -
          bearerAuth:

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsersUpdateRequest"
    get:
      summary: "Get user"
      tags:
        - "users"
      responses:
        200:
          description: "User detail"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUserResponse"
        404:
          description: "Not found"
      security:
        -
          bearerAuth:
...
  /admin/users/{id}:
    get:
      summary: "Admin user detail"
      tags:
        - "admin/users"
      responses:
        200:
          description: "Detail"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetailResponse"
        404:
          description: "Not found"
      security:
        -
          bearerAuth:

    patch:
      summary: "Admin update user"
      tags:
        - "admin/users"
      responses:
        200:
          description: "Updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetailResponse"
      security:
        -
          bearerAuth:

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserProfileUpdate"
...
  /posts:
    post:
      summary: "Create post"
      tags:
        - "posts"
      responses:
        201:
          description: "Created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
      security:
        -
          bearerAuth:

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
    get:
      summary: "List posts"
      tags:
        - "posts"
      responses:
        200:
          description: "List"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPosts"
      security:
        -
          bearerAuth:
  /topics/{id}/posts:
    get:
      summary: "Posts by topic"
      tags:
        - "posts"
      responses:
        200:
          description: "List"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPosts"
      security:
        -
          bearerAuth:
...
  /community/report:
    post:
      summary: "Report content"
      tags:
        - "community"
      responses:
        200:
          description: "Reported"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"
                  id:
                    type: "string"
      security:
        -
          bearerAuth:

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityReport"
  /community/block:
    post:
      summary: "Block user"
      tags:
        - "community"
      responses:
        200:
          description: "Blocked"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  ok:
                    type: "boolean"
                  id:
                    type: "string"
      security:
        -
          bearerAuth:

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityBlock"
...
  /posts/{postId}:
    patch:
      summary: "게시글 수정"
      parameters:
        - in: "path"
          name: "postId"
          required: true
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
      responses:
        200:
          description: "OK"
    delete:
      summary: "게시글 삭제"
      parameters:
        - in: "path"
          name: "postId"
          required: true
          schema:
            type: "string"
      responses:
        204:
          description: "No Content"
...