# services/api/src/modules/users/users.controller.ts
import {
  Body,
  Controller,
  ForbiddenException,
  Get,
  NotFoundException,
  Param,
  Patch,
  Query,
} from '@nestjs/common';
import { ApiBearerAuth, ApiQuery, ApiTags } from '@nestjs/swagger';
import { UsersService } from './users.service';
import { UpdateUserDto } from './update-user.dto';
import { CurrentUser } from '../auth/current-user.decorator';

type CurrentRequestUser = {
  id?: string;
  activityAccountId?: string | null;
};

@ApiTags('users')
@ApiBearerAuth()
@Controller('users')
export class UsersController {
  constructor(private readonly users: UsersService) {}

  // 내 정보 조회 (JWT 토큰 기반)
  @Get('me')
  async getMe(@CurrentUser() user: CurrentRequestUser) {
    const userId = user?.id;
    const userData = userId ? await this.users.byId(userId) : null;
    if (!userData) throw new NotFoundException('User not found');
    return {
      ok: true,
      data: {
        ...this.serializeUser(userData),
        activityAccountId: user?.activityAccountId ?? null,
      },
    };
  }

  @ApiQuery({ name: 'q', required: false, type: String })
  @ApiQuery({ name: 'limit', required: false, type: Number })
  @Get('search')
  async search(
    @Query('q') q: string = '',
    @Query('limit') limit?: string,
  ) {
    const take = Math.min(50, Math.max(1, Number(limit ?? 20) || 20));
    const results = await this.users.search(q ?? '', take);

    const items = results.map((item) => ({
      id: item.id,
      displayName: item.displayName,
      nickname: item.profile?.nickname ?? null,
      headline: item.profile?.headline ?? null,
      bio: item.profile?.bio ?? null,
      avatarUri: item.profile?.avatarUri ?? null,
      region1: item.region1 ?? null,
      region2: item.region2 ?? null,
      interests: item.profile?.interests ?? [],
      targetAccountId: item.ownerBridge?.activityAccounts?.[0]?.id ?? null,
    }));

    return { ok: true, data: items, items };
  }

  @Patch(':id')
  async update(
    @Param('id') id: string,
    @CurrentUser() user: any,
    @Body() body: UpdateUserDto,
  ) {
    const authId = user?.sub ?? user?.id;
    if (!authId || authId !== id) {
      throw new ForbiddenException('You can only update your own profile');
    }

    const interests = this.normalizeInterests(body.interests);
    const marketingOptIn = this.normalizeBoolean(body.marketingOptIn);

    const bioValue = typeof body.bio === 'string' ? body.bio.trim() : undefined;
    const headlineValue = typeof body.headline === 'string' ? body.headline.trim() : undefined;
    const avatarUriValue = typeof body.avatarUri === 'string' ? body.avatarUri.trim() : undefined;
    
    const updated = await this.users.updateProfile(id, {
      displayName: body.displayName?.trim() || undefined,
      nickname: body.nickname?.trim() || undefined,
      bio: bioValue === undefined ? undefined : bioValue.length ? bioValue : null,
      region1: body.region1?.trim() || undefined,
      region2: body.region2?.trim() || undefined,
      interests,
      marketingOptIn,
      headline:
        headlineValue === undefined ? undefined : headlineValue.length ? headlineValue : null,
      avatarUri:
        avatarUriValue === undefined ? undefined : avatarUriValue.length ? avatarUriValue : null,
    });

    return { ok: true, data: this.serializeUser(updated) };
  }

  // 공용/모바일에서 단건 조회 시 사용 예시
  @Get(':id')
  async getUserById(@Param('id') id: string) {
    const user = await this.users.byPublicId(id);
    if (!user) throw new NotFoundException('User not found');
    return { ok: true, data: this.serializePublicUser(user) };
  }

  private serializePublicUser(user: any) {
    return {
      id: user.id,
      displayName: user.displayName ?? null,
      region1: user.region1 ?? null,
      region2: user.region2 ?? null,
      profile: user.profile
        ? {
            nickname: user.profile.nickname ?? null,
            bio: user.profile.bio ?? null,
            headline: user.profile.headline ?? null,
            avatarUri: user.profile.avatarUri ?? null,
            interests: user.profile.interests ?? [],
            badges: user.profile.badges ?? [],
          }
        : null,
    };
  }
  private serializeUser(user: any) {
    const visibility = user?.profile?.visibility as Record<string, any> | undefined;
    return {
      id: user.id,
      email: user.email,
      displayName: user.displayName ?? null,
      status: user.status,
      role: user.role,
      provider: user.provider,
      region1: user.region1 ?? null,
      region2: user.region2 ?? null,
      pointsBalance: user.pointsBalance ?? 0,
      createdAt: user.createdAt,
      profile: user.profile
        ? {
            nickname: user.profile.nickname,
            bio: user.profile.bio,
            headline: user.profile.headline ?? null,
            avatarUri: user.profile.avatarUri ?? null,
            interests: user.profile.interests ?? [],
            badges: user.profile.badges ?? [],
            marketingOptIn: visibility?.marketingOptIn ?? false,
          }
        : null,
      counts: user._count
        ? {
            posts: user._count.posts ?? 0,
            reports: user._count.reportsToMe ?? 0,
            messages: user._count.messages ?? 0,
          }
        : undefined,
    };
  }

  private normalizeInterests(value: UpdateUserDto['interests']) {
    if (Array.isArray(value)) {
      return value
        .map((item) => (typeof item === 'string' ? item.trim() : ''))
        .filter((item) => item.length > 0);
    }

    if (typeof value === 'string') {
      return value
        .split(',')
        .map((item) => item.trim())
        .filter((item) => item.length > 0);
    }

    return undefined;
  }

  private normalizeBoolean(value: UpdateUserDto['marketingOptIn']) {
    if (value === undefined) return undefined;
    if (typeof value === 'boolean') return value;
    if (typeof value === 'string') {
      return ['true', '1', 'yes', 'y'].includes(value.toLowerCase());
    }
    return undefined;
  }
}